backend/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.3
flask-cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
SQLAlchemy==2.0.20
typing_extensions
Werkzeug==2.3.7
pyrad==2.4
netaddr==0.8.0
Flask-Migrate==4.1.0
alembic==1.16.4
psycopg2-binary
Flask-APScheduler==1.12.4
python-dotenv==0.21.0
